Added express.router and split out routes to a set of new files
